refactor(filid): split cache-manager into caches organ (T1-A)
Split core/infra/cache-manager/cache-manager.ts (383 lines, 20 exports
across 6 cache concerns) into an organ-first structure:

- caches/prompt-context-cache.ts  — prompt context read/write/has
- caches/session-cache.ts         — session tracking + GC (cross-concern
                                     filesystem knowledge — see DETAIL.md)
- caches/boundary-cache.ts        — boundary read/write
- caches/fractal-map-cache.ts     — fractal map read/write/remove + FractalMap type
- caches/run-hash-cache.ts        — run hash save/get
- caches/guide-cache.ts           — guide injected flag
- cache-manager.ts                — slim facade (12 lines with export type)

session-cache intentionally knows filename conventions of prompt-context,
guide, boundary, and fmap caches (required for pruneOldSessions and
removeSessionFiles). This cross-concern coupling is documented in DETAIL.md
as an intentional constraint.

- src/index.ts / src/core/index.ts: empty diff (facade path unchanged)
- cache-manager.ts: 383 -> 12 lines
- INTENT.md: 30 lines (<=50)
- bundle delta: max +4.03% (context-injector.mjs, well under 5% gate)
